pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "chardet",
    "docx>=0.2.4",
    "faiss-cpu",
    "langchain-community>=0.3.1",
    "langchain-openai>=0.2.1",
    "langchain>=0.3.1",
    "openai>=1.50.2",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.9",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "requests>=2.32.3",
    "scikit-learn==1.5.2",
    "streamlit>=1.38.0",
    "tiktoken",
    "trafilatura>=1.12.2",
]
